refactor(api-gateway): split apiKeyValidation.ts per provider
apiKeyValidation.ts was at 383/400 lines — adding the next provider
validator would have crossed the ESLint max-lines error threshold.
PR 3's planned /settings tts UX work is likely to add another validator
or two, so the split lands here as prerequisite work.

New layout:
  utils/
    apiKeyValidation.ts          (72 lines — dispatcher only)
    apiKeyValidation.test.ts     (4 dispatcher routing tests)
    apiKeyValidation/
      types.ts                   (shared types + VALIDATION_MESSAGES)
      openrouter.ts + .test.ts   (validateOpenRouterKey)
      elevenlabs.ts + .test.ts   (validateElevenLabsKey + permissions parser)
      mistral.ts + .test.ts      (validateMistralKey)
      zaiCoding.ts + .test.ts    (validateZaiCodingKey)

Public surface unchanged: callers continue to import { validateApiKey,
ApiKeyValidationResult, validateXxxKey } from './apiKeyValidation.js'.
The dispatcher re-exports per-provider validators so test files that
exercise a single provider still work, but application code uses the
dispatcher.

Per-provider tests now colocate with their validators (per
.claude/rules/02-code-standards.md). 35 tests pass across the 5 files
(was 35 in 1 file).

From PR #959 round 4 (claude-review). Closes the absorption block of
TTS-epic backlog items; PR 3 main scope (/settings tts UX, gateway
TtsConfig CRUD) starts on a clean foundation with no half-resolved
review nits.
